Major update: support for keyword tracking, logging deatiled stats, volume plots, city-based filtering, integration

+++ data/interactionfollower/packages/io/io.py
@@ -4,6 +4,7 @@ import sys
 import time
 import logging
 import os
+from pathlib import Path
 
 class Io:
 
@@ -30,10 +31,12 @@ class Io:
         self.json_write_file = json_write_file
         self.json_read_file  = json_read_file
         self.c_saved         = 0
+        self.daily_c_saved   = 0
         self.stime           = time.time()
-        self.ids             = self.get_userids()
+        self.inputs          = self.get_input()
+        self.cities          = {}
 
-    def _read_userids(self):
+    def _read_csv(self):
         with open(self.json_read_file, "r") as f:
             ids = list(f.read().splitlines())
             logging.debug("IO {0}: user ids successfully read for the input file {0}. They are:".format(self.ID, self.json_read_file))
@@ -41,40 +44,66 @@ class Io:
         return ids
             
     def instantiate_file(self, filepath):
-        with open(filepath, 'w+') as f: 
-            pass
+        if not Path(filepath).is_file():
+            self.daily_c_saved = 0
+            with open(filepath, 'w+') as f: 
+                pass
 
-    def get_userids(self):
+    def get_input(self):
         """
-        reads userids from a file and creates a list object
+        reads lines from a file and creates a list object
         """
         logging.info("Reading user_ids.")
         try:
-            self.ids = list(set(self._read_userids()))
-            return self.ids
+            self.inputs = list(set(self._read_csv()))
+            return self.inputs
         except Exception as ex:
             logging.error("Error while reading user_ids: %s",repr(ex))
             logging.info("Exiting program.")
             exit()
 
     def update(self):
-        ids = self._read_userids()
-        if set(ids) == set(self.ids):
+        ids = self._read_csv()
+        if set(ids) == set(self.inputs):
             return False
-        self.ids = list(set(ids))
+        self.inputs = list(set(ids))
         return True
 
-    def save(self, data, jsonfilename):
+    def save_status(self, data, jsonfilename):
         for at_i in range(3):        
             try: 
                 f = open(jsonfilename, "a", encoding='utf-8', newline='')
-                f.write(data)
+                f.write((str(data._json)))
+                f.write('\n')
                 self.c_saved += 1
+                self.daily_c_saved += 1
                 logging.debug('Tweet saved, total count: {}'.format(self.c_saved))
                 return True
             except Exception as ex:
                 logging.error("Save was unsuccessful:", ex)
                 time.sleep(0.5)
         return False
+
+    def save_uid(self, userid, W):
+        fname = '{0}.csv'.format(W)
+        if W not in self.cities.keys:
+            self.cities[W] = 1
+        else:
+            self.cities[W] += 1
+            
+        for at_i in range(3):        
+            try: 
+                f = open(fname, "a", encoding='utf-8', newline='')
+                f.write(userid)
+                f.write('\n')
+                logging.info('UserID {0} saved to file: {1}'.format(userid, fname))
+                return True
+            except Exception as ex:
+                logging.error("Save was unsuccessful:", ex)
+                time.sleep(0.5)
+        return False
+
+    def set_filename(self, filename):
+        self.jsonfilename = filename
